User Logic implmentation
